requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
black==24.8.0
blinker==1.8.2
Brotli==1.1.0
cffi==1.17.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
frozenlist==1.4.1
identify==2.6.0
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
pycares==4.4.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
tzdata==2024.1
virtualenv==20.26.3
Werkzeug==3.0.4
yarl==1.9.8
